requirements.txt
flask==2.3.2
bcrypt==4.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
flask-httpauth==4.8.0
